add useReferenceMany Hooks
